+++ tests/Helpers/HelpersTest.php
@@ -4,19 +4,11 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\CoreBundle\Tests\Helpers;
 
-use Dbp\Relay\CoreBundle\Helpers\GuzzleTools;
 use Dbp\Relay\CoreBundle\Helpers\MimeTools;
 use PHPUnit\Framework\TestCase;
-use Psr\Log\NullLogger;
 
 class HelpersTest extends TestCase
 {
-    public function testGuzzleTools()
-    {
-        $middleware = GuzzleTools::createLoggerMiddleware(new NullLogger());
-        $this->assertNotNull($middleware);
-    }
-
     public function testMimeTools()
     {
         $this->assertSame(MimeTools::getDataURI('foobar', 'text/html'), 'data:text/html;base64,Zm9vYmFy');