+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v0.1.203
+
 * Add Rest/Query/Parameters:getBool helper function and tests
 * Removed long deprecated GuzzleTools class
 * Removed long deprecated Helpers\Locale class